move raw file operations from FileSystemPlayer to FileSystemProvider
